pyproject.toml
[tool.poetry]
name = "hw-service-applied-python-with-db"
version = "0.1.0"
description = ""
authors = ["Nikita Filonenko"]
packages = [
    { include = "app", from = "." }
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.115.12"
pydantic = "^2.11.1"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.40"
redis = "^5.2.1"
python-jose = "^3.4.0"
passlib = "^1.7.4"
fastapi-users = {extras = ["sqlalchemy"], version = "^14.0.1"}
python-dotenv = "^1.1.0"
alembic = "^1.15.2"
aioredis = "^2.0.1"
prometheus-fastapi-instrumentator = "^7.1.0"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
apscheduler = "^3.11.0"
pydantic-settings = "^2.8.1"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pylint = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
